+++ package.json
@@ -40,35 +40,35 @@
     "organization": true
   },
   "devDependencies": {
-    "@types/jest": "^29.5.3",
+    "@types/jest": "^29.5.4",
     "@types/node": "^18",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "aws-cdk": "^2.87.0",
-    "esbuild": "^0.18.11",
-    "eslint": "^8.44.0",
-    "eslint-config-prettier": "^8.8.0",
+    "@typescript-eslint/eslint-plugin": "^6.4.1",
+    "aws-cdk": "^2.99.0",
+    "esbuild": "^0.19.2",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-header": "^3.1.1",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^5.0.0",
-    "jest": "^29.6.1",
-    "jest-extended": "^4.0.0",
+    "jest": "^29.6.4",
+    "jest-extended": "^4.0.1",
     "jest-junit": "^16.0.0",
     "npm-check-updates": "^16",
-    "projen": "^0.71.133",
+    "projen": "^0.72.31",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
+    "typescript": "~5.1.6"
   },
   "dependencies": {
-    "@aws-cdk/aws-lambda-python-alpha": "^2.87.0-alpha.0",
-    "@aws-cdk/aws-servicecatalogappregistry-alpha": "^2.87.0-alpha.0",
-    "@aws-sdk/client-dynamodb": "^3.369.0",
-    "@aws-sdk/client-ec2": "^3.369.0",
-    "@aws-sdk/client-rds": "^3.369.0",
-    "@aws-sdk/client-ssm": "^3.369.0",
-    "@aws-solutions-constructs/aws-lambda-dynamodb": "^2.41.0",
-    "aws-cdk-lib": "^2.87.0",
-    "cdk-nag": "^2.27.67",
+    "@aws-cdk/aws-lambda-python-alpha": "^2.99.0-alpha.0",
+    "@aws-cdk/aws-servicecatalogappregistry-alpha": "^2.99.0-alpha.0",
+    "@aws-sdk/client-dynamodb": "^3.398.0",
+    "@aws-sdk/client-ec2": "^3.399.0",
+    "@aws-sdk/client-rds": "^3.399.0",
+    "@aws-sdk/client-ssm": "^3.398.0",
+    "@aws-solutions-constructs/aws-lambda-dynamodb": "^2.42.0",
+    "aws-cdk-lib": "^2.99.0",
+    "cdk-nag": "^2.27.114",
     "constructs": "^10.0.5",
     "source-map-support": "^0.5.21"
   },
@@ -76,6 +76,6 @@
     "node": ">= 18.0.0"
   },
   "license": "Apache-2.0",
-  "version": "1.5.1",
+  "version": "1.5.2",
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
 }
