+++ source/cli/pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "instance_scheduler_cli"
-version = "1.5.1"
+version = "1.5.2"
 description = "Instance Scheduler on AWS CLI"
 license = "Apache-2.0"
 authors = ["Amazon Web Services"]
@@ -23,14 +23,16 @@ boto3 = "^1.26.90"
 jmespath = "^1.0.1"
 
 [tool.poetry.group.dev.dependencies]
-black = "^23.7.0"
+black = "^23.9.1"
 boto3-stubs-lite = {extras = ["cloudformation", "lambda"], version = "^1.26.90"}
 cli-test-helpers = "^3.4.0"
-flake8 = "^6.0.0"
+flake8 = "^6.1.0"
 isort = "^5.12.0"
-moto = {extras = ["cloudformation", "lambda"], version = "^4.1.12"}
-mypy = "^1.4.1"
-pytest = "^7.4.0"
+# held back, 4.18.0 is a breaking change
+jsonschema = "~4.17.3"
+moto = {extras = ["cloudformation", "lambda"], version = "^4.2.4"}
+mypy = "^1.5.1"
+pytest = "^7.4.2"
 pytest-cov = "^4.1.0"
-types-jmespath = "^1.0.2.6"
-types-requests = "^2.31.0.1"
+types-jmespath = "^1.0.1"
+types-requests = "^2.31.0.6"
